Pulled daemonset creation out to its own package
