+++ cmd/operator/main.go
@@ -1,15 +1,16 @@
 package operator
 
 import (
-	"fmt"
+	"kube-image-prefetch/internal/prefetcher"
+	"os"
 	"time"
 
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
+	"k8s.io/client-go/tools/clientcmd"
 )
 
 const (
@@ -33,21 +34,11 @@ func Run() error {
 
 		images, pullSecrets := parseDeployments(deployments)
 
-		ds, _ := clientset.AppsV1().DaemonSets(namespace).Get(name, metav1.GetOptions{})
-		if ds == nil || ds.ObjectMeta.Name == "" {
-			ds = buildDaemonSet()
-		}
+		ds := prefetcher.BuildDaemonset(images, pullSecrets)
 
-		ds.Spec.Template.Spec.Containers = []corev1.Container{}
-		ds.Spec.Template.Spec.ImagePullSecrets = pullSecrets
-
-		i := 0
-		for img := range images {
-			ds.Spec.Template.Spec.Containers = append(
-				ds.Spec.Template.Spec.Containers,
-				buildPrefetchContainer(img, i),
-			)
-			i++
+		existingDs, err := clientset.AppsV1().DaemonSets(namespace).Get(name, metav1.GetOptions{})
+		if err == nil {
+			ds.ResourceVersion = existingDs.ResourceVersion
 		}
 
 		if ds.ResourceVersion == "" {
@@ -66,7 +57,11 @@ func Run() error {
 func getClient() (*kubernetes.Clientset, error) {
 	config, err := rest.InClusterConfig()
 	if err != nil {
-		return nil, err
+		kubeconfigPath := os.Getenv("KUBECONFIG")
+		if kubeconfigPath == "" {
+			kubeconfigPath = os.Getenv("HOME") + "/.kube/config"
+		}
+		config, err = clientcmd.BuildConfigFromFlags("", kubeconfigPath)
 	}
 
 	return kubernetes.NewForConfig(config)
@@ -81,80 +76,20 @@ func getDeployments(clientset *kubernetes.Clientset) ([]appsv1.Deployment, error
 	return dps.Items, nil
 }
 
-func parseDeployments(deployments []appsv1.Deployment) (map[string]bool, []corev1.LocalObjectReference) {
-	images := map[string]bool{}
+func parseDeployments(deployments []appsv1.Deployment) ([]string, []corev1.LocalObjectReference) {
+	imagesMap := map[string]bool{}
 	pullSecrets := []corev1.LocalObjectReference{}
 	for _, dp := range deployments {
 		for _, container := range append(dp.Spec.Template.Spec.Containers, dp.Spec.Template.Spec.InitContainers...) {
-			images[container.Image] = true
+			imagesMap[container.Image] = true
 		}
 		pullSecrets = append(pullSecrets, dp.Spec.Template.Spec.ImagePullSecrets...)
 	}
 
+	images := []string{}
+	for img := range imagesMap {
+		images = append(images, img)
+	}
+
 	return images, pullSecrets
-}
-
-func buildDaemonSet() *appsv1.DaemonSet {
-	labels := map[string]string{
-		"app": name,
-	}
-	return &appsv1.DaemonSet{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:   name,
-			Labels: labels,
-		},
-		Spec: appsv1.DaemonSetSpec{
-			Selector: &metav1.LabelSelector{
-				MatchLabels: labels,
-			},
-			Template: corev1.PodTemplateSpec{
-				ObjectMeta: metav1.ObjectMeta{
-					Labels: labels,
-				},
-				Spec: corev1.PodSpec{
-					InitContainers: []corev1.Container{{
-						Name:            "init",
-						Image:           image,
-						ImagePullPolicy: corev1.PullAlways,
-						Args: []string{
-							"-command", "copy",
-							"-dest", "/mount/sleep",
-						},
-						VolumeMounts: []corev1.VolumeMount{{
-							Name:      "share",
-							MountPath: "/mount",
-						}},
-					}},
-					Containers:       []corev1.Container{},
-					ImagePullSecrets: []corev1.LocalObjectReference{},
-					Volumes: []corev1.Volume{{
-						Name: "share",
-						VolumeSource: corev1.VolumeSource{
-							EmptyDir: &corev1.EmptyDirVolumeSource{},
-						},
-					}},
-				},
-			},
-		},
-	}
-}
-
-func buildPrefetchContainer(img string, index int) corev1.Container {
-	return corev1.Container{
-		Name:            fmt.Sprintf("prefetch-%d", index),
-		Image:           img,
-		ImagePullPolicy: corev1.PullIfNotPresent,
-		Command:         []string{"/mount/sleep"},
-		Args:            []string{"-command", "sleep"},
-		VolumeMounts: []corev1.VolumeMount{{
-			Name:      "share",
-			MountPath: "/mount",
-		}},
-		Resources: corev1.ResourceRequirements{
-			Limits: corev1.ResourceList{
-				corev1.ResourceCPU:    resource.MustParse("1m"),
-				corev1.ResourceMemory: resource.MustParse("10M"),
-			},
-		},
-	}
 }
